Created middleware handler file
