+++ README.md
@@ -7,3 +7,4 @@ Beginning of code was based off of Nic Jackson's building microservices in Go : 
 - Refactor to have a game related item instead of random product example
 - Tests for http handlers
 - Make validate take an interface instead of a specific struct
+- Improve setup documentation
