+++ handlers/products.go
@@ -1,8 +1,6 @@
 package handlers
 
 import (
-	"context"
-	"fmt"
 	"log"
 	"net/http"
 	"strconv"
@@ -60,32 +58,3 @@ func (p *Products) GetProducts(w http.ResponseWriter, request *http.Request) {
 }
 
 type KeyProduct struct{}
-
-// Validation middleware
-func (p *Products) MiddlewareProductValidation(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, request *http.Request) {
-		product := &data.Product{}
-
-		err := product.FromProductJSON(request.Body)
-		if err != nil {
-			p.l.Println("[ERROR] deserializing product", err)
-			http.Error(w, "Error reading product", http.StatusBadRequest)
-			return
-		}
-
-		// validate the product
-		err = product.ValidateProduct()
-		if err != nil {
-			p.l.Println("[ERROR] validating product", err)
-			http.Error(w, fmt.Sprintf("Error validating product: %s", err), http.StatusBadRequest)
-			return
-		}
-
-		// Add the product to the context
-		context := context.WithValue(request.Context(), KeyProduct{}, product)
-		newRequest := request.WithContext(context)
-
-		// Call the next handler, which can be another middleware or the final handler
-		next.ServeHTTP(w, newRequest)
-	})
-}
